Resolved issues. QuickFix.
