+++ buyer/tests/test_serializers.py
@@ -1,36 +1,10 @@
 """Serializers testing module to check correct adding data through them"""
 import pytest
 
-from buyer.models import Buyer, Offer
 from buyer.serializers.buyer_serializers import BuyerSerializer
 from buyer.serializers.offer_serializers import OfferSerializer
-from car.tests.test_serializers import car, engine
-
-
-@pytest.fixture
-def buyer():
-    """Fixture to add buyer instance."""
-    buyer = Buyer.objects.create(
-        full_name="F I O",
-        age=50,
-        gender="male",
-        balance=1111,
-        is_active=True,
-    )
-    return buyer
-
-
-@pytest.fixture
-def offer(car, buyer):
-    """Fixture to add offer instance."""
-    offer = Offer.objects.create(
-        buyer=buyer,
-        car=car,
-        max_cost=5637,
-        quantity=1,
-        is_active=True,
-    )
-    return offer
+from buyer.tests.buyer_app_fixtures import buyer, offer
+from car.tests.car_app_fixtures import car, engine
 
 
 @pytest.mark.django_db
